The controller is created. Controller connected to entry point
